+++ http/stats.go
@@ -29,7 +29,7 @@ import (
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/io/election"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/io/sender"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/man"
-	plWorker "gitlab.jiagouyun.com/cloudcare-tools/datakit/pipeline/worker"
+	plstats "gitlab.jiagouyun.com/cloudcare-tools/datakit/pipeline/stats"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs"
 )
 
@@ -82,22 +82,22 @@ type DatakitStats struct {
 
 	AvailableInputs []string `json:"available_inputs"`
 
-	HostName     string `json:"hostname"`
-	Version      string `json:"version"`
-	BuildAt      string `json:"build_at"`
-	Branch       string `json:"branch"`
-	Uptime       string `json:"uptime"`
-	OSArch       string `json:"os_arch"`
-	IOChanStat   string `json:"io_chan_stats"`
-	PLWorkerStat string `json:"pl_wroker_stats"`
-	Elected      string `json:"elected"`
-	Cgroup       string `json:"cgroup"`
-	CSS          string `json:"-"`
+	HostName   string `json:"hostname"`
+	Version    string `json:"version"`
+	BuildAt    string `json:"build_at"`
+	Branch     string `json:"branch"`
+	Uptime     string `json:"uptime"`
+	OSArch     string `json:"os_arch"`
+	IOChanStat string `json:"io_chan_stats"`
+	Elected    string `json:"elected"`
+	Cgroup     string `json:"cgroup"`
+	CSS        string `json:"-"`
 
-	InputsStats map[string]*io.InputsStat `json:"inputs_status"`
-	SenderStat  map[string]*sender.Metric `json:"sender_stat"`
-	IoStats     io.IoStat                 `json:"io_stats"`
-	HTTPMetrics map[string]*apiStat       `json:"http_metrics"`
+	InputsStats map[string]*io.InputsStat  `json:"inputs_status"`
+	SenderStat  map[string]*sender.Metric  `json:"sender_stat"`
+	IoStats     io.IoStat                  `json:"io_stats"`
+	PLStats     []plstats.ScriptStatsROnly `json:"pl_stats"`
+	HTTPMetrics map[string]*apiStat        `json:"http_metrics"`
 
 	WithinDocker bool            `json:"docker"`
 	AutoUpdate   bool            `json:"auto_update"`
@@ -119,7 +119,6 @@ var (
 - 系统类型   : {{.OSArch}}
 - 容器运行   : {{.WithinDocker}}
 - IO 消耗统计: {{.IOChanStat}}
-- Pipeline Worker 统计: {{.PLWorkerStat}}
 - 自动更新   ：{{.AutoUpdate}}
 - 选举状态   ：{{.Elected}}
 	`
@@ -321,7 +320,7 @@ func GetStats() (*DatakitStats, error) {
 		WithinDocker:   datakit.Docker,
 		IOChanStat:     io.ChanStat(),
 		IoStats:        io.GetIoStats(),
-		PLWorkerStat:   plWorker.ShowPLWkrStats().String(),
+		PLStats:        plstats.ReadStats(),
 		Elected:        fmt.Sprintf("%s::%s|%s", ns, elected, who),
 		Cgroup:         cgroup.Info(),
 		AutoUpdate:     datakit.AutoUpdate,