+++ pipeline/funcs/fn_cast_test.go
@@ -7,6 +7,7 @@ package funcs
 
 import (
 	"testing"
+	"time"
 
 	tu "gitlab.jiagouyun.com/cloudcare-tools/cliutils/testutil"
 )
@@ -130,12 +131,17 @@ func TestCast(t *testing.T) {
 				}
 				return
 			}
-
-			if err := runner.Run(tc.in); err != nil {
-				t.Error(err)
+			ret, err := runner.Run("test", map[string]string{},
+				map[string]interface{}{
+					"message": tc.in,
+				}, "message", time.Now())
+			if err != nil {
+				t.Fatal(err)
+			}
+			if ret.Error != nil {
+				t.Error(ret.Error)
 				return
 			}
-			ret := runner.Result()
 			t.Log(ret)
 			v := ret.Fields[tc.outkey]
 			tu.Equals(t, tc.expect, v)