+++ pipeline/funcs/fn_strfmt_test.go
@@ -7,6 +7,7 @@ package funcs
 
 import (
 	"testing"
+	"time"
 
 	tu "gitlab.jiagouyun.com/cloudcare-tools/cliutils/testutil"
 )
@@ -119,16 +120,20 @@ strfmt(bb, "%v%s", a.forth, "tone")
 				}
 				return
 			}
-
-			err = runner.Run(tc.in)
+			ret, err := runner.Run("test", map[string]string{},
+				map[string]interface{}{
+					"message": tc.in,
+				}, "message", time.Now())
 			if err != nil {
+				t.Fatal(err)
+			}
+			if ret.Error != nil {
 				if tc.fail {
 					t.Logf("[%d]expect error: %s", idx, err)
 				} else {
 					t.Error(err)
 				}
 			} else {
-				ret := runner.Result()
 				t.Log(ret)
 				v := ret.Fields[tc.outKey]
 				tu.Equals(t, tc.expected, v)