+++ plugins/externals/logfwd/logfwd.go
@@ -150,6 +150,7 @@ func startTailing(lg *logging, fn tailer.ForwardFunc, stop <-chan struct{}) {
 		MultilineMatch:        lg.MultilineMatch,
 		RemoveAnsiEscapeCodes: lg.RemoveAnsiEscapeCodes,
 		ForwardFunc:           fn,
+		FromBeginning:         true,
 	}
 
 	tailer, err := tailer.NewTailer(lg.LogFiles, opt, lg.Ignore)